Reorganizing the source code.  This should allow us to start reorganizing the tests (splitting the test_api.py file apart, etc).
Do not explicitly add "inventory" prefix to the logger names.  This is
no longer needed due to the reorganization of the code.
