+++ inventory/api/metrics.py
