+++ inventory/app/common.py
