+++ inventory/app/events.py
