+++ inventory/app/utils.py
