+++ manage.py
@@ -1,7 +1,7 @@
 import os
 from flask_script import Manager  # class for handling a set of commands
 from flask_migrate import Migrate, MigrateCommand
-from app import db, create_app
+from inventory.app import db, create_app
 
 
 app = create_app(config_name=os.getenv('APP_SETTINGS'))
